refactor(auth): 优化验证码发送及登录流程
- 将验证码发送由RabbitMQ消息队列改为Feign远程调用
- 登录接口增加调用用户服务Feign客户端获取用户详细信息
- 去除过时注释及多余代码，改进异常处理机制
- 调整验证码Redis缓存有效期存储方式，统一为5分钟
- 重命名Captcha模块邮件消费者为控制器，改用REST接口提供邮件验证码发送
- 整合Feign客户端调用Captcha服务发送验证码，增强模块解耦
- 优化代码日志输出，增强调试信息可读性
- 修改Activity和Club控制器路径统一为/api/club
- 更新多模块application.yml配置，修复RabbitMQ和Redis账号密码
- 新增聊天服务模块，包含会话、消息管理及长轮询支持
- 添加聊天服务相关实体、Mapper和服务实现，支持消息发送与会话管理
- 新增用户服务配置及过滤器，完善请求上下文传递机制
- 调整项目模块结构，重命名Interaction服务为Chat服务并添加异步配置
- 移除部分无用注释和类，代码结构更清晰规范

+++ services/resource-service/src/main/java/org/example/entity/Resource.java
@@ -11,9 +11,8 @@ import java.time.LocalDateTime;
 @TableName("resource")
 public class Resource {
     @TableId(type = IdType.AUTO)
-    private Integer ResourcesId;
-
-    private Integer ownerId;
+    private Long ResourcesId;
+    private Long ownerId;
     private String imagePath;
     private String title;
     private String content;
